requirements.txt
autopep8==1.6.0
et-xmlfile==1.1.0
mysql-connector-python==8.0.28
numpy==1.22.2
openpyxl==3.0.9
pandas==1.4.1
protobuf==3.19.4
pycodestyle==2.8.0
python-dateutil==2.8.2
pytz==2021.3
six==1.16.0
toml==0.10.2
